// src/components/profile/types.ts
export interface UserInfo {
  maNguoiDung: number;
  tenDangNhap: string;
  email: string;
  hoTen: string;
  soDienThoai: string;
  ngaySinh: string;
  gioiTinh: string;
  diaChi: string;
  tienSuBenh?: string;
  diUng?: string;
  maBenhNhan: number;
}
// ...
export interface ProfileInfoCardProps {
  userInfo: UserInfo;
  isEditingProfile: boolean;
  setIsEditingProfile: (val: boolean) => void;
  setUserInfo: React.Dispatch<React.SetStateAction<UserInfo | null>>;
  handleUpdateProfile: () => void;
}